+++ .gitignore
@@ -17,3 +17,4 @@
 .DS_Store
 *.rsa
 *.pub
+rsa/
