+++ configs/server.go
@@ -1,4 +1,4 @@
-package models
+package configs
 
 type Server struct {
 	Port string `json:"port"`
