+++ main.go
@@ -11,8 +11,6 @@ import (
 )
 
 func main() {
-	tools.StartServerDouble()
-
 	dataBase.NewDBConnection()
 
 	port, err := tools.GetServerPort("configs/server.json")
