+++ tools/response.go
@@ -1,0 +1,5 @@
+package tools
+
+type Response struct {
+	Data string `json:"data"`
+}
