+++ tools/token.go
@@ -1,0 +1,5 @@
+package tools
+
+type Token struct {
+	Token string `json:"token"`
+}
